# docs/swagger.yaml
basePath: /api/v1
definitions:
  response.Data:
    properties:
      code:
        $ref: '#/definitions/response.ResCode'
      data: {}
      msg: {}
    type: object
...
  schema.UserCommentCreateRequest:
    properties:
      content:
        description: Comment content | 评论内容
        example: 很有见地的评论
        minLength: 1
        type: string
      parent_id:
        description: Parent comment ID (used when replying to comment) | 父评论ID（回复评论时使用）
        example: 1
        type: integer
      post_id:
        description: Post ID | 帖子ID
        example: 1
        type: integer
      reply_to_user_id:
        description: Reply target user ID (used when replying to user) | 回复目标用户ID（回复用户时使用）
        example: 2
        type: integer
    required:
    - content
    - post_id
    type: object
...
  schema.UserCommentListResponse:
    properties:
      list:
        description: Comment list | 评论列表
        items:
          $ref: '#/definitions/schema.UserCommentListItem'
        type: array
      page:
        description: Current page number | 当前页码
        type: integer
      page_size:
        description: Items per page | 每页数量
        type: integer
      total:
        description: Total count | 总数量
        type: integer
    type: object
  schema.UserCommentUpdateRequest:
    properties:
      content:
        description: Comment content | 评论内容
        example: 更新后的评论内容
        minLength: 1
        type: string
      id:
        description: Comment ID | 评论ID
        example: 1
        type: integer
    required:
    - content
    - id
    type: object
...
  schema.UserUpdateAvatarRequest:
    properties:
      avatar_url:
        description: Avatar URL | 头像URL
        example: https://example.com/avatar.jpg
        type: string
    required:
    - avatar_url
    type: object
...
host: localhost:9876
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: PokeForum is a forum application based on Gin framework | PokeForum
    是一个基于 Gin 框架的论坛应用程序
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PokeForum API
  version: "1.0"
paths:
...
  /comments:
    get:
      consumes:
      - application/json
      description: Get paginated comment list for specified post with sorting support
        | 分页获取指定帖子的评论列表,支持排序
      parameters:
      - description: Post ID | 帖子ID
        in: query
        name: post_id
        required: true
        type: integer
      - description: Page number | 页码
        in: query
        name: page
        required: true
        type: integer
      - description: Items per page | 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      - description: 'Sort field: created_at, like_count | 排序字段:created_at, like_count'
        example: '"created_at"'
        in: query
        name: sort_by
        type: string
      - description: Is descending order | 是否降序
        in: query
        name: sort_desc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved successfully | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserCommentListResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server internal error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get comment list | 获取评论列表
      tags:
      - '[User]Comments | [用户]评论'
    post:
      consumes:
      - application/json
      description: User creates new comment, supports replying to comments and users
        | 用户创建新评论,支持回复评论和回复用户
      parameters:
      - description: Create comment request | 创建评论请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserCommentCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created successfully | 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserCommentCreateResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server internal error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Create comment | 创建评论
      tags:
      - '[User]Comments | [用户]评论'
    put:
      consumes:
      - application/json
      description: User updates their own comment content | 用户更新自己的评论内容
      parameters:
      - description: Update comment request | 更新评论请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserCommentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated successfully | 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserCommentUpdateResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "403":
          description: Insufficient permissions | 权限不足
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server internal error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Update comment | 更新评论
      tags:
      - '[User]Comments | [用户]评论'
...
  /posts:
    get:
      consumes:
      - application/json
      description: Get post list with pagination and sorting support. Supports filtering
        by category (via ID or slug) and keyword search on title. Pinned posts are
        returned separately in pinned_posts field | 获取帖子列表,支持分页和排序。支持通过版块ID或slug筛选,以及标题关键词搜索。置顶帖子单独返回在pinned_posts字段中
      parameters:
      - description: Category ID | 版块ID
        in: query
        name: category_id
        type: integer
      - description: Category slug | 版块slug
        in: query
        name: slug
        type: string
      - description: Keyword for title search | 标题关键词搜索
        in: query
        name: keyword
        type: string
      - default: 1
        description: Page number, default 1 | 页码,默认1
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page, default 20, max 100 | 每页数量,默认20,最大100
        in: query
        name: page_size
        type: integer
      - default: latest
        description: 'Sort method: latest (newest), hot (popular), essence (featured)
          | 排序方式:latest(最新)、hot(热门)、essence(精华)'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved successfully | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserPostListResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get post list | 获取帖子列表
      tags:
      - '[User]Topic Posts | [用户]主题贴'
    post:
      consumes:
      - application/json
      description: User publishes a new topic post | 用户发布新的主题帖
      parameters:
      - description: Post information | 帖子信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserPostCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Published successfully | 发布成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserPostCreateResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Not logged in | 未登录
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Publish new post | 发布新帖
      tags:
      - '[User]Topic Posts | [用户]主题贴'
    put:
      consumes:
      - application/json
      description: User edits their own post (can be operated once every three minutes)
        | 用户编辑自己的帖子(每三分钟可操作一次)
      parameters:
      - description: Post information | 帖子信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserPostUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Edited successfully | 编辑成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserPostUpdateResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Not logged in | 未登录
          schema:
            $ref: '#/definitions/response.Data'
        "403":
          description: Insufficient permissions or too frequent operations | 权限不足或操作过于频繁
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Edit post | 编辑帖子
      tags:
      - '[User]Topic Posts | [用户]主题贴'
...
  /profile/avatar:
    put:
      consumes:
      - application/json
      description: Update current logged-in user's avatar | 修改当前登录用户的头像
      parameters:
      - description: Update avatar request | 修改头像请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserUpdateAvatarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update successful | 修改成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserUpdateAvatarResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Internal server error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Update avatar | 修改头像
      tags:
      - '[User]Profile | [用户]个人中心'
...
  /profile/comments:
    get:
      consumes:
      - application/json
      description: Get comments published by specified user, supports pagination,
        retrieves current logged-in user's comments if user_id not provided | 获取指定用户发布的评论列表，支持分页，不传user_id则获取当前登录用户的评论
      parameters:
      - description: User ID, queries current logged-in user if not provided | 用户ID，不传则查询当前登录用户
        in: query
        name: user_id
        type: integer
      - description: Page number | 页码
        in: query
        name: page
        required: true
        type: integer
      - description: Items per page | 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Retrieve successful | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserProfileCommentsResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Internal server error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get user comments list | 获取用户评论列表
      tags:
      - '[User]Profile | [用户]个人中心'
  /profile/email/verify:
    post:
      consumes:
      - application/json
      description: Verify user email authenticity through verification code | 通过验证码验证用户邮箱真实性
      parameters:
      - description: Verify email request | 验证邮箱请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.EmailVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Verification successful | 验证成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.EmailVerifyResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "404":
          description: Verification code does not exist or has expired | 验证码不存在或已过期
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Internal server error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Verify email | 验证邮箱
      tags:
      - '[User]Profile | [用户]个人中心'
  /profile/email/verify-code:
    post:
      consumes:
      - application/json
      description: Send verification code to user's registered email for email verification
        | 向用户注册邮箱发送验证码，用于邮箱验证
      produces:
      - application/json
      responses:
        "200":
          description: Send successful | 发送成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.EmailVerifyCodeResponse'
              type: object
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "429":
          description: Too many requests | 发送频率过高
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Internal server error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Send email verification code | 发送邮箱验证码
      tags:
      - '[User]Profile | [用户]个人中心'
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"